support for macro release and reworked macro flags
that unknown variable for macros was actually an index for the release
value. all macro flags are stored in the very beginning of each macro
table. the end flag is now only an end flag, and there is no more
data to be put after it. previously it was used to store the loop index.